+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Raster Vision 0.10.0
 Features
 ^^^^^^^^^^^^
 
+- Add confusion matrix to eval for semantic segmentation `#786 <https://github.com/azavea/raster-vision/pull/786>`__
 - Handle "ignore" class for semantic segmentation `#783 <https://github.com/azavea/raster-vision/pull/783>`__
 
 Bug Fixes